spit identity (WIP)
