+++ service/src/services/mod.rs
@@ -5,6 +5,8 @@ mod identity_service_session;
 
 mod ep_health;
 pub(in crate::services) use self::ep_health::*;
+mod ep_configure;
+pub(in crate::services) use self::ep_configure::*;
 mod ep_search_identity;
 pub(in crate::services) use self::ep_search_identity::*;
 mod ep_generate_user_name;
